Refactor omnigraph db module layout
